Optional Annotation for some ZenMethods
Usually something with more than one Input/Output (unless it's an Array)
should have an @'Optional annotation. This is because you do not have to
type out every single parameter with "null" or "0", making the code
cleaner.

I am kind of unsure whether or not the Optional annotation for the
AdditiveArray is fitting because I'm not sure whether how essential the
Additives are to the Arc Furnace. I can remove it if you think it does
not fit :)